Refactor BytecodeTrie to its own module
